perf-tests/requirements.txt
PyYAML==5.1.2
requests==2.22.0
kubernetes==8.0.1
pytest==4.4.1
ipaddress==1.0.22 # >= 1.0.17
cffi==1.12.3
certifi==2019.6.16
urllib3==1.25.3
pytest-html==2.1.1
pytest-repeat==0.8.0
locust==1.1.1